ripple/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
anthropic>=0.39.0
openai>=1.30.0
pyyaml>=6.0
sse-starlette>=2.0.0
pgvector>=0.3.0
